+++ src/Rector/SetList.php
@@ -4,15 +4,15 @@ namespace Wieni\wmcodestyle\Rector;
 
 final class SetList
 {
-    public const CODE_QUALITY = __DIR__ . '/../../rector/code-quality.php';
+    public const CODE_QUALITY = __DIR__ . '/../../rector/sets/code-quality.php';
 
-    public const CODING_STYLE = __DIR__ . '/../../rector/coding-style.php';
+    public const CODING_STYLE = __DIR__ . '/../../rector/sets/coding-style.php';
 
-    public const DEPENDENCY_INJECTION = __DIR__ . '/../../rector/dependency-injection.php';
+    public const DEPENDENCY_INJECTION = __DIR__ . '/../../rector/sets/dependency-injection.php';
 
-    public const EARLY_RETURN = __DIR__ . '/../../rector/early-return.php';
+    public const EARLY_RETURN = __DIR__ . '/../../rector/sets/early-return.php';
 
-    public const HOOK_EVENT_DISPATCHER = __DIR__ . '/../../rector/hook-event-dispatcher.php';
+    public const HOOK_EVENT_DISPATCHER = __DIR__ . '/../../rector/sets/hook-event-dispatcher.php';
 
-    public const TYPE_DECLARATION = __DIR__ . '/../../rector/type-declaration.php';
+    public const TYPE_DECLARATION = __DIR__ . '/../../rector/sets/type-declaration.php';
 }